git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/branches/feature/rendermanager@28145 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ bin/autogen.sh
@@ -15,5 +15,7 @@ done
 
 autoconf $@
 rc=$?
-rm -f $MACROFILE
+if [[ $rc ]] ; then
+  rm -f $MACROFILE
+fi
 exit $rc
